pyproject.toml
[project]
name = "linkedinyearreview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "html2image>=2.0.7",
    "humanize>=4.14.0",
    "jinja2>=3.1.6",
    "numerize>=0.12",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "playwright>=1.57.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "streamlit>=1.52.2",
]
